test: add a test for unit rri algo

+++ include/rri_algo.h
@@ -9,7 +9,7 @@
 #include <gmp.h>
 
 // is expected to be of size at least 2 deg(P)
-int subdiv_algo_ext(fmpz_poly_t in_poly, float sol[], float start, float end,
+int subdiv_algo_ext(fmpz_poly_t in_poly, double sol[], double start, double end,
                     int *next_index_p);
 
 #endif // RRI_ALGO_H
